Remove Go name-resolution nonsequitor.
This reference doesn't make sense in its current spot, as that name
resolution is done by the frontend server. It will be worth re-adding
this information to the section on `fetch`, after the move to the
Transformation API.

Also, fix the paths to the main go files in the README, and add a
disclaimer to the top.

Change-Id: Ida41007af056d730708e50151099a7d5b5d4c83b