Default to fetching from the sign URL.
Change the code to allow omission of URLSet.Fetch from the config, in
which case the content will be fetched from the sign URL, and the fetch
URL must not be specified. Updated the example config to reflect this
default. This is a safer default, as it prevents URL A from being
fetched and signed as URL B (modulo a network-level attack, such as a
MITM of google.com or the original URL).

Also made some backwards-incompatible changes to the config:
- Moved SamePath into URLSet.Fetch, since it only makes sense if that is
  defined.
- Moved ErrorOnStatefulHeaders into URLSet.Sign, since that is useful
  even in the typical case.

Also introduced a new URLSet.Fetch.DomainRE field, to allow the frontend
to specify a particular edge node to fetch from.