Implement transactional semantics and refactor.
Add a second read/isExpired check inside the write lock to ensure the
read-modify-write is atomic.

Extracted ocsp and ocspMu out of CertCache into a new InMemory
implementation of Updateable, and made a Chained Updateable that
implements the cascading from memory to disk.

Merged maybeUpdateOCSP into readOCSP, to eliminate the possibility of
skew between ocspFile and ocspUpdateAfter.

Rename some methods and add comments.