Add a carriage return to the start of a <pre> tag ONLY if there is already a newline at the start.
This is necessary for idempotency when re-parsing. The HTML spec dops the first newline, so when there are multiple newlines in sequence, each subsequent parse will differ.

For example, before this CL:
  <pre>&#13;&#13;blah</pre>
becomes
  <pre>&#13;blah</pre>
and if re-parsed becomes
  <pre>blah</pre>

With the change in this CL:
  <pre>&#13;&#13;blah</pre>
stays the same and can be re-parsed over and over.

However, what remains unchanged is if there is only a SINGLE newline at the start of <pre>.

So
  <pre>&#13;blah</pre>
becomes
  <pre>blah</pre>
and if re-parsed stays as
  <pre>blah</pre>

PiperOrigin-RevId: 213713024