Fix dom traversal bug after extra <title> tags were removed.
Changed logic to only remove the "current" node, as opposed to look ahead removal. The edge case that came up was:

<head><title>a</title><title>b</title>unseen text node</head>

The for loop would set "next" to the b title, then perform logic on the a title. The logic would then look ahead and remove b. Returning to the recursion stack would still have "next" pointing to the now orphaned node and therefore all the siblings after would be unseen.

PiperOrigin-RevId: 215465577