Export SelectVersion() function, to be called before Process(). As soon as I tried to integrate this into the packager, I realized I need to do them separately. Imagine a request:
AMP-Cache-Transform: google;v="2", any;v="1"

If only version 1 is supported, then the A-C-T parser needs to determine that,
and select `any` instead of `google` (without running the transforms multiple
times).

Incremented the minimum version to 1, since this now relies on a sentinel value
of 0 to mean "not specified" in the proto3 message.

Addresses #29.

PiperOrigin-RevId: 218195708