Add DNS preconnects when different origin is detected.
Had to refactor the code so that the first pass initially tokenizes the URLs, then the second pass over the tokens actually converts to cache URL and adds the preconnects, if necessary. The original implementation hid the url details in sub-libraries, where it was difficult to determine if preconnect was necessary or not.

PiperOrigin-RevId: 223527873