Reduce initial Go heap from 768M to 128M.
The 128M number was arbitrary, and some tuning may be necessary to find
an optimal number.

This ends up triggering golang/go#29355, and thus requires the
workaround suggested in that thread: construct a flyweight TypedArray
each type the JS wants to mutate the slice.

The workaround slows things down around 18% from ~39ms per document to
~46ms, but I have an idea for a workaround for _that_ (TODO in the
code).

Delete the Go1.12 port because it's not worth maintaining it at this
early prototypal stage.