Remove warning about code, and add one about spec.
Remove the warning about the code being in developer preview; we
consider it stable enough to deploy in production now. Both code and
specification are still evolving, but this is handled by the requirement
to keep the packager updated per the `AMP-Cache-Transform` version.

Add a warning about SXG security considerations. This is an attempt to
distill advice from the risks outlined in the spec.