Be explicit about the SXG cache lifetime. (#323)
The previous behavior didn't specify a max age, which leads to
unspecified behavior.

Set to "max-age=0" for now, which allows caching on intermediaries, but
recommends they update every time.

A future change could relax that max age. However, be careful when doing
so, as the failure behavior of an expired SXG is a different UX than
that of an expired unsigned response.