Travis: GO111MODULES=on, disable Go 1.10
GO111MODULES=on allows it to be tested with the exact version of the
dependencies specified in go.mod for Go 1.11 and Go.12, and not the
newest version of each such dep.

This fixes a build failure caused by a commit to
github.com/stretchr/testify 3 days ago; see
https://travis-ci.org/github/ampproject/amppackager/builds/710522601.

Go 1.10 did not support modules in any form. It did support vendored
directories, but I'm not sure how to make Travis not run `go get` and
update them. So, let's finally stop testing on Go 1.10.
https://travis-ci.org/github/ampproject/amppackager/builds/710522601.