Fix bug with URL rewriting on elements with style attribute
The inline style attribute is one of the first things checked for URLs. And we unfortunately skipped all the rest of the URL rewriting when there is one.

This also exposed a bug with rewriting an image with `src=#`. We would generate a new srcset for the image, but the validator will not accept `# 123w` as a valid source in the srcset. Now, we just skip such srcs.

PiperOrigin-RevId: 326745297