Update Prometheus metrics (#543)
This is an overhaul of the Prometheus metrics. The goal is to improve
usability by following best practices.

Use Histograms rather than summaries.
* Summaries can not be aggregated, this makes them unusable when you run
multiple instances of the service.
* Histograms provide the ability to use "heatmap" views of the data.

Drop counters that duplicate data we already have in the histogram `_count` metric.

Rename metrics to follow Prometheus naming conventions. For example,
prefixing custom metrics with the app name as a `Namespace`.

Related to https://github.com/ampproject/amppackager/issues/357

Signed-off-by: SuperQ <superq@gmail.com>