reorder <meta name=viewport> to be always the 2nd item
We want the <meta name=viewport> to be as early as possible in the <head> tag.
It is necessary for this tag to be before the dvh polyfill because the <meta name=viewport> can trigger a change on the innerHeight of the document on iOS which doesn't cause a resize event. This can cause the calculation in the dvh polyfill script to be incorrect.

PiperOrigin-RevId: 427491146