Sync from Google (#670)
* Fix AMP Story-related "cta-image" attributes when set to "none".

A valid value is "none" (which should hide the link icon). However, the transformer rewrites the value "none" to a relative URL when configured for doing so.

If the value is "none", it shouldn't get replaced with the absolute url.

PiperOrigin-RevId: 473286485

* Followup: Fix AMP Story-related "cta-image" attributes when set to "none".

A valid value is "none" (which should hide the link icon). However, the transformer rewrites the value "none" to a relative URL when configured for doing so.

If the value is "none", it shouldn't get replaced with the absolute url.

PiperOrigin-RevId: 474009294

Co-authored-by: Googler <noreply@google.com>